backend/requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
gunicorn==20.1.0
asgiref==3.5.2
pytz==2022.6
sqlparse==0.4.3
django-rest-framework==0.1.0
djangorestframework==3.13.1